requirements.txt
fastapi
uvicorn[standard]
torch
git+https://github.com/openai/CLIP.git
pillow
numpy
aiohttp
pydantic
